pyproject.toml
[project]
name = "gemini-discord"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "google>=3.0.0",
    "google-genai>=1.52.0",
    "python-dotenv>=1.2.1",
]
